Removed unused report parameter config classes. Slight refactoring of saved report configurations.
