+++ docker/config.env
@@ -1,3 +1,5 @@
 CERT_IMPORT_DIRECTORY=/certificates
 aquariusServiceUser=PUT_AQUARIUS_USERNAME_HERE
 aquariusServiceEndpoint=PUT_AQUARIUS_ENDPOINT_HERE_WITHOUT_/AQUARIUS/
+S3_BUCKET=aqcu-report-configs-test
+S3_ENDPOINT=http://localhost:8081
