+++ src/main/java/gov/usgs/aqcu/controller/LookupController.java
@@ -3,8 +3,6 @@ package gov.usgs.aqcu.controller;
 import java.util.List;
 import java.util.Map;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import gov.usgs.aqcu.parameter.FieldVisitDatesRequestParameters;
 import gov.usgs.aqcu.parameter.FindInDerivationChainRequestParameters;
 import gov.usgs.aqcu.parameter.GetUpchainRatingModelsRequestParameters;
@@ -18,7 +16,6 @@ import gov.usgs.aqcu.model.lookup.TimeSeriesBasicData;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.WebRequest;
@@ -106,16 +103,6 @@ public class LookupController {
 	@GetMapping(value="/units", produces={MediaType.APPLICATION_JSON_VALUE})
 	public ResponseEntity<?> getUnits() throws Exception {
 		return new ResponseEntity<List<String>>(lookupsService.getUnits(), new HttpHeaders(), HttpStatus.OK);
-	}
-
-	@GetMapping(value="/reportParameterConfig/{reportType}", produces={MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<?> getReportParameterConfig(@PathVariable String reportType) throws Exception {
-		return new ResponseEntity<String>(new ObjectMapper().writeValueAsString(lookupsService.getReportParameterConfig(reportType)), new HttpHeaders(), HttpStatus.OK);
-	}
-		
-	@GetMapping(value="/report/types", produces={MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<?> getReportTypes() throws Exception {
-		return new ResponseEntity<String>(lookupsService.getReportTypes(), new HttpHeaders(), HttpStatus.OK);
 	}
 
 	/*
