+++ src/main/java/gov/usgs/aqcu/model/config/GroupConfig.java
@@ -3,6 +3,9 @@ package gov.usgs.aqcu.model.config;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class GroupConfig {
 	private List<String> authorizedUsers;
