--- src/main/java/gov/usgs/aqcu/retrieval/ReportParameterConfigLookupService.java
@@ -1,19 +1,0 @@
-package gov.usgs.aqcu.retrieval;
-
-import org.springframework.stereotype.Repository;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-
-import gov.usgs.aqcu.model.report.ReportParameterConfig;
-
-@Repository
-public class ReportParameterConfigLookupService {
-
-	public ReportParameterConfig getByReportType(String reportType){
-		return ReportParameterConfig.getByReportType(reportType);
-	}
-	
-	public String getReportTypes() throws JsonProcessingException{
-		return ReportParameterConfig.getReportTypes();
-	}
-}
